+++ CUDA/InputTests/InputTest.cpp
@@ -36,7 +36,7 @@ void InputTest::saveDoubleTestVectorToXML(const vector<double> &p_vecDoubleValue
 	if(p_vecDoubleValues.size() > 0)
 	{
 		const vector<AttributeMapping> &vecAttributeMappings = m_pParentTestSet->getAttributeMappings();
-		unsigned uAttributesMappingSize = vecAttributeMappings.size();
+		size_t uAttributesMappingSize = vecAttributeMappings.size();
 		for(unsigned uAttributeIndex = 0;uAttributeIndex < uAttributesMappingSize;++uAttributeIndex)
 		{
 			const AttributeMapping &attributeMappingData = vecAttributeMappings[uAttributeIndex];
